Fix parsing of historical data
The measurement unit is different - millions of m3, and on top of that, when the client requests the page the
numeric values are somewhat raw, not matching what you see on the browser. They are more precise and less human readable.
Comments were included about that as well to make debugging easier