Implement spider for the new AESA website
The new AESA website exposes an endpoint responding with a JSON file with the measurement data, removing the need for HTML scraping.
A new service method was created to consume that endpoint. The others will be removed, no need to maintain it anymore.
Guava introduced for collection utilities (Lists.newArrayList etc), RestTemplate configuration included to make Jackson work with that
endpoint (unfortunately they respond with the wrong content type). More cleanup to follow